Implement new Climber
The new climber system was implemented, with a solenoid for the arms, and two motors for legDown and legDrive. legDown has analog down motion and fixed percentage up motion, on the right trigger and bumper for weapons respectively, and the legDrive motor has analog on the left weapons trigger. The arm deploy is bound to the weapons window button, with weapons menu and a TBD driver button also needing to be pressed as confirmation.